:recycle: [태진] refactor: 컴포넌트 형식 통일
