+++ src/components/Modal/ModalStyled.jsx
@@ -1,44 +1,8 @@
 import styled from 'styled-components';
 import COLORS from '../../utils/colors';
 import FONTS from '../../utils/Fonts';
-import { formatDate } from '../../utils/formatDate';
-import Buttons from '../Buttons';
 
-const Modal = ({ openModal, handleClose, profileImageURL, sender, relationship, content, createdAt }) => {
-  if (!openModal) {
-    return <></>;
-  }
-
-  const handleDeemClick = () => {
-    handleClose();
-  };
-
-  return (
-    <>
-      <Deem onClick={handleDeemClick} />
-      <Container>
-        <CardHeader>
-          <CardProfile profileImageURL={profileImageURL} />
-          <div>
-            <Sender>
-              From. <b>{sender}</b>
-            </Sender>
-            <Relationship>{relationship}</Relationship>
-          </div>
-          <Date>{formatDate(createdAt)}</Date>
-        </CardHeader>
-        <CardContents dangerouslySetInnerHTML={{ __html: content }} />
-        <CardFooter>
-          <div onClick={handleClose}>
-            <Buttons buttonType="Primary40">확인</Buttons>
-          </div>
-        </CardFooter>
-      </Container>
-    </>
-  );
-};
-
-const Deem = styled.div`
+export const Deem = styled.div`
   position: fixed;
   top: 0;
   left: 0;
@@ -48,7 +12,7 @@ const Deem = styled.div`
   background: rgba(0, 0, 0, 0.6);
 `;
 
-const Container = styled.div`
+export const Container = styled.div`
   width: 600px;
   height: 476px;
   border-radius: 16px;
@@ -68,7 +32,7 @@ const Container = styled.div`
   }
 `;
 
-const CardHeader = styled.div`
+export const CardHeader = styled.div`
   display: flex;
   align-items: center;
   gap: 2rem;
@@ -76,7 +40,7 @@ const CardHeader = styled.div`
   border-bottom: 1px solid ${COLORS.gray200};
 `;
 
-const CardContents = styled.div`
+export const CardContents = styled.div`
   padding-top: 20px;
   height: 280px;
   overflow: scroll;
@@ -92,14 +56,14 @@ const CardContents = styled.div`
   }
 `;
 
-const CardFooter = styled.div`
+export const CardFooter = styled.div`
   display: flex;
   justify-content: center;
   //   align-items: center;
   padding-top: 10px;
 `;
 
-const CardProfile = styled.div`
+export const CardProfile = styled.div`
   width: 5.6rem;
   height: 5.6rem;
   border-radius: 10rem;
@@ -111,7 +75,7 @@ const CardProfile = styled.div`
   background-position: center;
 `;
 
-const Relationship = styled.span`
+export const Relationship = styled.span`
   font-size: 1.4rem;
   color: ${COLORS.blue500};
   background-color: ${COLORS.blue100};
@@ -119,15 +83,13 @@ const Relationship = styled.span`
   border-radius: 0.4rem;
 `;
 
-const Sender = styled.p`
+export const Sender = styled.p`
   font-size: 2rem;
   margin-bottom: 0.5rem;
 `;
 
-const Date = styled.div`
+export const Date = styled.div`
   margin-left: auto;
   color: ${COLORS.gray400};
   ${FONTS.font14_Regular}
 `;
-
-export default Modal;
