+++ src/components/Navbar/Navbar.jsx
@@ -1,18 +1,17 @@
-import styled from 'styled-components';
-import COLORS from '../utils/colors';
 import { Link, useLocation } from 'react-router-dom';
-import logoImg from '../assets/rolling_Logo.svg';
-import Layout from '../layout/Layout';
-import Buttons from './Buttons';
+import logoImg from '../../assets/rolling_Logo.svg';
+import Layout from '../../layout/Layout';
+import Buttons from '../Buttons';
+import * as S from './NavbarStyled';
 
 const Navbar = () => {
   const currentLocation = useLocation();
 
   return (
-    <Nav>
+    <S.Nav>
       <Layout>
-        <Container>
-          <Ul>
+        <S.Container>
+          <S.Ul>
             <li>
               <Link to="/">
                 <h1>
@@ -27,29 +26,11 @@ const Navbar = () => {
                 </Link>
               )}
             </li>
-          </Ul>
-        </Container>
+          </S.Ul>
+        </S.Container>
       </Layout>
-    </Nav>
+    </S.Nav>
   );
 };
-
-const Nav = styled.nav`
-  border-bottom: 1px solid ${COLORS.gray200};
-  position: relative;
-  z-index: 10;
-  background-color: ${COLORS.white};
-`;
-
-const Container = styled.div`
-  height: 64px;
-`;
-
-const Ul = styled.ul`
-  height: 100%;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-`;
 
 export default Navbar;
