+++ src/components/Portal/index.js
@@ -1,0 +1,1 @@
+export * from './ModalPortal';
