+++ src/components/Reaction/ReactionStyled.jsx
@@ -1,15 +1,7 @@
 import styled from 'styled-components';
 import COLORS from '../../utils/colors';
 
-const Reaction = ({ reaction, containerStyle }) => {
-  return (
-    <Container style={containerStyle}>
-      {reaction.emoji} {reaction.count}
-    </Container>
-  );
-};
-
-const Container = styled.li`
+export const Container = styled.li`
   background-color: ${COLORS.black};
   background: rgba(0, 0, 0, 0.54);
   color: #fff;
@@ -21,5 +13,3 @@ const Container = styled.li`
     font-size: 1.2rem;
   }
 `;
-
-export default Reaction;
