+++ src/components/ReactionAddButton/ReactionAddButton.jsx
@@ -1,10 +1,9 @@
 import EmojiPicker from 'emoji-picker-react';
-import styled from 'styled-components';
 import unionImg from '../../assets/icons/Union.svg';
-import COLORS from '../../utils/colors';
 import { useParams } from 'react-router-dom';
 import { createReaction } from '../../api/api';
 import { usePopOver } from '../../hooks/usePopOver';
+import * as S from './ReactionAddButtonStyled';
 
 const ReactionAddButton = ({ triggerUpdate }) => {
   const { id } = useParams();
@@ -28,10 +27,10 @@ const ReactionAddButton = ({ triggerUpdate }) => {
 
   return (
     <>
-      <AddBtn onClick={handleReactionAdd}>
+      <S.AddBtn onClick={handleReactionAdd}>
         <img src={unionImg} alt="personIcon" />
         추가
-      </AddBtn>
+      </S.AddBtn>
       <div ref={ref}>
         <EmojiPicker
           open={isOpen}
@@ -42,20 +41,5 @@ const ReactionAddButton = ({ triggerUpdate }) => {
     </>
   );
 };
-
-const AddBtn = styled.button`
-  border: 1px solid ${COLORS.gray300};
-  padding: 6px 16px;
-  display: flex;
-  font-size: 1.6rem;
-  gap: 5px;
-  align-items: center;
-  border-radius: 6px;
-  font-weight: 500;
-  @media (max-width: 960px) {
-    padding: 3px 8px;
-    font-size: 1.2rem;
-  }
-`;
 
 export default ReactionAddButton;
