+++ src/components/ReactionAddButton/ReactionAddButtonStyled.jsx
@@ -1,0 +1,17 @@
+import styled from 'styled-components';
+import COLORS from '../../utils/colors';
+
+export const AddBtn = styled.button`
+  border: 1px solid ${COLORS.gray300};
+  padding: 6px 16px;
+  display: flex;
+  font-size: 1.6rem;
+  gap: 5px;
+  align-items: center;
+  border-radius: 6px;
+  font-weight: 500;
+  @media (max-width: 960px) {
+    padding: 3px 8px;
+    font-size: 1.2rem;
+  }
+`;
