+++ src/pages/Paper/components/MessageCard/index.js
@@ -1,0 +1,1 @@
+export { default } from './MessageCard';
